Auth Using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -64,6 +65,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)){
+            abort(403, 'Access denied!');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -76,6 +80,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)){
+            abort(403, 'Access denied!');
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your question has been updated!!');
     }
@@ -88,6 +95,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)){
+            abort(403, 'Access denied!');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted!');
     }
